Added interactive cd function.
Added interactive cd function which lists directories of current directory with a number in front.
User can choose to switch to dir by only typing in the number in front of the directory. Function reruns after successful cd to ask again if user wants to switch to a subdir.
Entering q cancels the function.